Update ss.ca.R
Added color configurability. 
In the the other functions the col list is ss.col in this one it is f.colours. Just noting that difference.
Elements 1 - 5 (f.colours) were left as previously defined in the function. 
Element 6 background color
Element 7 is the label text color
Element 8 is the values color